+++ R/pdf_pages.R
@@ -8,7 +8,7 @@
 #'              or "all", "even", or "odd" to include all pages, just the even pages,
 #'              or just the odd pages.
 #' @examples
-#' f <- pdf_blank(length = 8L)
+#' f <- pdf_create_blank(length = 8L)
 #' pdf_pages(f, pages = 1:4)
 #' pdf_pages(f, pages = -(1:4))
 #' pdf_pages(f, pages = "all")