+++ tests/testthat/test-pdf_pad_length.R
@@ -1,14 +1,14 @@
-test_that("`pdf_blank()`", {
+test_that("`pdf_create_blank()`", {
     on.exit(rm_temp_pdfs(), add = TRUE)
 
-    f <- pdf_blank(length = 4L)
+    f <- pdf_create_blank(length = 4L)
     expect_equal(qpdf::pdf_length(f), 4L)
 })
 
 test_that("`pdf_pad_length()`", {
     on.exit(rm_temp_pdfs(), add = TRUE)
 
-    f <- pdf_blank(length = 1L)
+    f <- pdf_create_blank(length = 1L)
     expect_equal(qpdf::pdf_length(f), 1L)
 
     f2 <- pdf_pad_length(f, minimum = 8L, multiples_of = 4L)