+++ tests/testthat/test-pnpmisc.R
@@ -3,7 +3,7 @@ test_that("miscellaneous functions", {
 
     skip_if(tools::find_gs_cmd()[[1L]] == "")
 
-    f1 <- pdf_blank()
+    f1 <- pdf_create_blank()
     f2 <- pdf_gs(f1)
 
     expect_equal(qpdf::pdf_length(f2), 1L)
@@ -30,7 +30,7 @@ test_that("miscellaneous functions", {
 test_that("`pdf_pages()`", {
     on.exit(rm_temp_pdfs(), add = TRUE)
 
-    f <- pdf_blank(length = 8L)
+    f <- pdf_create_blank(length = 8L)
     expect_equal(pdf_pages(f, pages = 1:4), 1:4)
     expect_equal(pdf_pages(f, pages = -(1:4)), 5:8)
     expect_equal(pdf_pages(f, pages = "all"), 1:8)